# Reduce launch overhead by generating code to encode arguments.