Before:

```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  11.041 μs … 197.375 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     16.750 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.016 μs ±  11.460 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  ▅ ▁▁▃█▂
  █▅█████▄▃▃▄▄▃▃▄▃▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▂
  11 μs           Histogram: frequency by time         74.4 μs <

 Memory estimate: 784 bytes, allocs estimate: 32.
```

After:

```
BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):   5.584 μs … 452.500 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):      6.333 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   20.696 μs ±  31.537 μs  ┊ GC (mean ± σ):  0.00% ± 0.00%

  █▅▄▂▁                                       ▁▂▁▁▁▁ ▁▂▂▁▁▁    ▁
  ██████▇▅▆▆▅▅▅▅▅▄▁▁▅▁▃▃▅▅▃▅▃▄▃▄▃▁▃▄▅▄▅▅▆▅▆▇██████████████████ █
  5.58 μs       Histogram: log(frequency) by time       105 μs <

 Memory estimate: 576 bytes, allocs estimate: 20.
```

For reference, CUDA.jl:

```
BenchmarkTools.Trial: 10000 samples with 10 evaluations.
 Range (min … max):  1.401 μs …  3.715 μs  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     1.470 μs              ┊ GC (median):    0.00%
 Time  (mean ± σ):   1.468 μs ± 46.336 ns  ┊ GC (mean ± σ):  0.00% ± 0.00%

   ▂▃▃▄▄▃▂▂▁▂▃▃▄▄▄▆▇▇█▇▄▃▂▁       ▁▁▁▁                       ▂
  ▇███████████████████████████▇███████▆▆▆▆▇▆▄▆▆▅▄▃▅▄▄▄▁▁▄▄▅▅ █
  1.4 μs       Histogram: log(frequency) by time     1.61 μs <

 Memory estimate: 64 bytes, allocs estimate: 1.
```

So a long way to go, but it's a solid improvement nonetheless.